move directory to utils
