+++ src/compiler/semantics/mod.rs
@@ -5,7 +5,7 @@ pub mod semantic_cube;
 use std::fmt::{Debug, Display};
 use lazy_static::lazy_static;
 use crate::utils::lazy_counter::Counter;
-use crate::directory::{self, Dir};
+use crate::utils::directory::{self, Dir};
 use self::item::{Kind, Item};
 use super::{parser::ast::*, semantics::semantic_cube::resolve, lexer::Span};
